requirements.txt
requests==2.33.1
python-dotenv>=1.0.0
rich>=13.4.0
